% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_lookup.R
\name{col_lookup}
\alias{col_lookup}
\title{Look up a column name}
\usage{
col_lookup(col, from, to)
}
\arguments{
\item{col}{Column name}

\item{from}{Column source name. Either "template", "AWQMS", or "stations".}

\item{to}{Name of the column in the source you are looking up. Either "template", "AWQMS", or "stations".}
}
\description{
Retrieve a column name in AWQMS based on the column name in the template (or vice versa)
}
