% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mloc.R
\name{df_mloc}
\alias{df_mloc}
\title{Monitoring Locations dataframe}
\usage{
df_mloc()
}
\value{
dataframe
}
\description{
Retrieve a dataframe with column names for the Monitoring Locations worksheet
in Oregon DEQ's continuous data submission template xlsx file v2.03.
Returns one row of data with all values set to NA.
Includes internal staff columns: "Reachcode", "Measure", "LLID", "River Mile",
"Permanent Identifier", "Monitoring Location Status ID", and
"Monitoring Location Status Comment".
}
