% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eco2.R
\name{get_eco2}
\alias{get_eco2}
\title{Get Ecoregion Level II}
\usage{
get_eco2(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as
\code{\link[sf:st_crs]{sf::st_crs}}. Typically entered using the
numeric EPSG value. Accepts a vector.}
}
\value{
data frame from feature service
}
\description{
The function will query the United States Environmental Protection Agency
Shared Enterprise Geodata and Services (SEGS) web service service to determine
the Ecoregion Level II. The x and y coordinates (longitude and latitude)
are used to select a specific ecoregion.
}
\details{
The feature service can be accessed at \url{https://geodata.epa.gov/arcgis/rest/services/ORD/USEPA_Ecoregions_Level_III_and_IV/MapServer/11}.
}
\seealso{
\code{\link{get_eco2code}}, \code{\link{get_eco2name}}
}
