% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eco4.R
\name{get_eco4code}
\alias{get_eco4code}
\title{Get Ecoregion Level IV code}
\usage{
get_eco4code(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs].
Typically entered using the numeric EPSG value. Accepts a vector.}
}
\value{
Ecoregion Level IV code in character format
}
\description{
The function will query the United States Environmental Protection Agency
Shared Enterprise Geodata and Services (SEGS) web service service to determine
the Ecoregion Level IV code. The x and y coordinates (longitude and latitude)
are used to select a specific ecoregion.
}
\details{
The feature service can be accessed at https://geodata.epa.gov/arcgis/rest/services/ORD/USEPA_Ecoregions_Level_III_and_IV/MapServer/7
}
\seealso{
[get_eco4()], [get_eco4name()]
}
