% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc10.R
\name{get_huc10}
\alias{get_huc10}
\title{Get HUC10 info}
\usage{
get_huc10(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as
\code{\link[sf:st_crs]{sf::st_crs}}. Typically entered using
the numeric EPSG value. Accepts a vector.}
}
\value{
sf object or data frame columns for HUC10 and HUC10_Name
}
\description{
The function will query Oregon DEQ's WBD feature service to determine the HUC10 watershed code and name. The x and y coordinates (longitude and latitude) are
used to select a specific HUC10 from the feature service. The WBD version is included with NHDH_OR_931v220.
}
\details{
The feature service can be accessed at \url{https://arcgis.deq.state.or.us/arcgis/rest/services/WQ/WBD/MapServer/2}.
The feature service column Name is changed to HUC10_Name.
}
\seealso{
\code{\link{get_huc10code}}, \code{\link{get_huc10name}}
}
