% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc10.R
\name{get_huc10code}
\alias{get_huc10code}
\title{Get HUC10 code}
\usage{
get_huc10code(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs].
Typically entered using the numeric EPSG value. Accepts a vector.}
}
\value{
The HUC10 code as character format
}
\description{
The function will query Oregon DEQ's WBD feature service to determine the
HUC10 watershed code. The x and y coordinates (longitude and latitude) are
used to select a specific HUC10 from the feature service.
}
\seealso{
[get_huc10()], [get_huc10name()]
}
