% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc10.R
\name{get_huc10name}
\alias{get_huc10name}
\title{Get HUC10 name}
\usage{
get_huc10name(x, y)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}
}
\value{
The HUC8 name
}
\description{
The function will query Oregon DEQ's WBD feature service to determine
the HUC10 watershed name. The x and y coordinates (longitude and latitude) are
used to select a specific HUC10 from the feature service.
}
\seealso{
[get_huc10()], [get_huc10code()]
}
