% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc12.R
\name{get_huc12}
\alias{get_huc12}
\title{Get HUC12 info}
\usage{
get_huc12(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as
\code{\link[sf:st_crs]{sf::st_crs}}. Typically entered using
the numeric EPSG value. Accepts a vector.}
}
\value{
data frame columns for HUC12 and HUC12_Name
}
\description{
The function will query Oregon DEQ's WBD feature service to determine the HUC12
subwatershed code and name. The x and y coordinates (longitude and latitude) are
used to select a specific HUC12 from the feature service. The WBD version is
included with NHDH_OR_931v220.
}
\details{
The feature service can be accessed at \url{https://arcgis.deq.state.or.us/arcgis/rest/services/WQ/WBD/MapServer/3}.
The feature service column Name is changed to HUC12_Name.
}
\seealso{
\code{\link{get_huc12code}}, \code{\link{get_huc12name}}
}
