% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc12.R
\name{get_huc12name}
\alias{get_huc12name}
\title{Get HUC12 name}
\usage{
get_huc12name(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs].
Typically entered using the numeric EPSG value. Accepts a vector.}
}
\value{
The HUC12 name
}
\description{
The function will query Oregon DEQ's WBD feature service to determine
the HUC12 subwatershed name. The x and y coordinates (longitude and latitude) are
used to select a specific HUC12 from the feature service.
}
\seealso{
[get_huc12()], [get_huc12code()]
}
