% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc8.R
\name{get_huc8}
\alias{get_huc8}
\title{Get HUC8 info}
\usage{
get_huc8(x, y)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}
}
\value{
data frame columns for HUC8 and HUC8_Name
}
\description{
The function will query Oregon DEQ's WBD feature service to determine the HUC8 subbasin code and name. The x and y coordinates (longitude and latitude) are
used to select a specific HUC8 from the feature service. The WBD version is included with NHDH_OR_931v220.
}
\details{
The feature service can be accessed at https://arcgis.deq.state.or.us/arcgis/rest/services/WQ/WBD/MapServer/1
The feature service column Name is changed to HUC8_Name.
}
\seealso{
[get_huc8code()] and [get_huc8name()]
}
