% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc8.R
\name{get_huc8name}
\alias{get_huc8name}
\title{Get HUC8 name}
\usage{
get_huc8name(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs].
Typically entered using the numeric EPSG value. Accepts a vector.}
}
\value{
The HUC8 name
}
\description{
The function will query Oregon DEQ's WBD feature service to determine
the HUC8 subbasin name. The x and y coordinates (longitude and latitude) are
used to select a specific HUC8 from the feature service.
}
\seealso{
[get_huc8()], [get_huc8code()]
}
