% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measure.R
\name{get_measure}
\alias{get_measure}
\title{Get measure value along Oregon NHDH_OR_931v220.}
\usage{
get_measure(pid, x, y, crs, return_sf = FALSE)
}
\arguments{
\item{pid}{The NHD Permanent_Identifier value as a string. Required.}

\item{x}{The longitude in decimal degrees. Required.}

\item{y}{The latitude in decimal degrees. Required.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs].
Typically entered using the numeric EPSG value.}

\item{return_sf}{Boolean. A TRUE value will return the sf object with data
frame columns for GNIS_Name, Permanent_Identifier, ReachCode, Measure, Snap.Lat, and Snap.Long.
FALSE will return the measure value as a character. Default is FALSE.}
}
\value{
sf object with data frame columns for Permanent_Identifier, ReachCode, Measure, Snap.Lat, and Snap.Long.
}
\description{
The function will query Oregon DEQ's NHD REST service to determine the measure
value along a target NHD flowline. The Permanent Identifier (pid) is used to
select a specific reach from the REST service. The x and y coordinates are snapped
to the nearest vertex on the selected flowline and the measure value determined.
If return_sf=TRUE the sf object is projected into ESPG:4326 (WGS84) which is
the expected default for [leaflet].
}
\details{
This function is intended to be a helper function for [launch_map] although it
may be used independently as well. Compatible with piping.

The feature service can be accessed at https://arcgis.deq.state.or.us/arcgis/rest/services/WQ/NHDH_ORDEQ/MapServer/1
}
