% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhdplushr.R
\name{get_nhdplushr}
\alias{get_nhdplushr}
\title{Get NHDplus HR info}
\usage{
get_nhdplushr(x, y, crs, search_dist = 100)
}
\arguments{
\item{x}{The longitude in decimal degrees. Required. Accepts a vector.}

\item{y}{The latitude in decimal degrees. Required. Accepts a vector.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs]. Typically entered using the numeric EPSG value. Accepts a vector.}

\item{search_dist}{The maximum search distance around x and y to look for features. Measured in meters. Default is 100.}
}
\value{
sf object
}
\description{
The function will query the NHDplus High Resolution (HR) feature service and return a
dataframe of info for the flowline. x and y coordinates are
used to select the reach.
}
\details{
The function will query the NHDplus High Resolution (HR) REST service from USGS and return a
dataframe containing the flowline info. The supplied x and y coordinates
(longitude and latitude) are used to select the closest flowline record and
calculate the measure value. Only the closest flowline within the search distance is returned.
If there are no flowlines within the search distance the returned dataframe will
contain all NAs. If two or more flowline records are equal distance to x and y
only the first record will be returned.

The NDHplus High Resolution feature service is at https://hydro.nationalmap.gov/arcgis/rest/services/NHDPlus_HR/MapServer/2/
}
