% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mloc_example.R
\docType{data}
\name{mloc_example}
\alias{mloc_example}
\title{Example Monitoring Location data}
\format{
An object of class \code{data.frame} with 160 rows and 39 columns.
}
\usage{
data(mloc_example)
}
\description{
Data frame of Oregon DEQ ambient network monitoring locations
}
\examples{
mlocs <- odeqmloctools::mloc_example()

}
\keyword{datasets}
