% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplus_na.R.R
\name{nhdplushr_na}
\alias{nhdplushr_na}
\title{NHDplus HR NA dataframe}
\usage{
nhdplushr_na()
}
\description{
Return a dataframe with all the columns in the NHDplus HR REST service with NAs.
Columns 'OBJECTID', 'Shape', and 'Shape_Length' are not included.
Columns 'Measure', 'Snap_Lat', 'Snap_Long', and 'Snap_Distance' are added for downstram binding.
}
\details{
More info at at https://hydro.nationalmap.gov/arcgis/rest/services/NHDPlus_HR/MapServer/2.
}
