% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orhuc10.R
\docType{data}
\name{orhuc10}
\alias{orhuc10}
\title{Oregon HUC10 Watersheds}
\format{
An object of class \code{data.frame} with 620 rows and 3 columns.
}
\usage{
data(orhuc10)
}
\description{
Table of Oregon ten digit hydrological codes and names.
Codes and names correspond to the Watershed Boundary Dataset (WBD) as published by
the U.S. Natural Resources and Conservation Service and U.S. Geological Survey
8/29/2013 and within National Hydrography Dataset (NHD) version NHDH_OR_931v220.
}
\details{
\itemize{
\item HUC10: Ten digit watershed code
\item HUC10_Name: Watershed name
\item State: States within the HUC
}
}
\examples{
huc10 <- odeqmloctools::orhuc10
}
\keyword{datasets}
