% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orhuc6.R
\docType{data}
\name{orhuc6}
\alias{orhuc6}
\title{Oregon HUC6 Basins}
\format{
An object of class \code{data.frame} with 15 rows and 3 columns.
}
\usage{
data(orhuc6)
}
\description{
Table of Oregon six digit hydrological codes and names.
Codes and names correspond to the Watershed Boundary Dataset (WBD) as published by
the U.S. Natural Resources and Conservation Service and U.S. Geological Survey
8/29/2013 and within National Hydrography Dataset (NHD) version NHDH_OR_931v220.
}
\details{
\itemize{
\item HUC6: Six digit basin code
\item HUC6_Name: Basin name
\item State: States within the HUC
}
}
\examples{
huc6 <- odeqmloctools::orhuc6
}
\keyword{datasets}
