% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orhuc6_12.R
\docType{data}
\name{orhuc6_12}
\alias{orhuc6_12}
\title{Oregon HUC6 - HUC12 Lookup}
\format{
An object of class \code{data.frame} with 3127 rows and 8 columns.
}
\usage{
data(orhuc6_12)
}
\description{
Lookup table of Oregon hydrological codes and names (HUC6 - HUC12).
Codes and names correspond to the Watershed Boundary Dataset (WBD) as published by
the U.S. Natural Resources and Conservation Service and U.S. Geological Survey
8/29/2013 and within National Hydrography Dataset (NHD) version NHDH_OR_931v220.
}
\details{
\itemize{
\item HUC6: Six digit basin code
\item HUC6_Name: Basin name
\item HUC8: Eight digit subbasin code
\item HUC8_Name: Subbasin name
\item HUC10: Ten digit watershed code
\item HUC10_Name: Watershed name
\item HUC12: Twelve digit subwatershed code
\item HUC12_Name: Subwatershed name
}
}
\examples{
huc6_12 <- odeqmloctools::orhuc6_12
}
\keyword{datasets}
