% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orhuc8.R
\docType{data}
\name{orhuc8}
\alias{orhuc8}
\title{Oregon HUC8 Subbasins}
\format{
An object of class \code{data.frame} with 91 rows and 3 columns.
}
\usage{
data(orhuc8)
}
\description{
Table of Oregon eight digit hydrological codes and names.
Codes and names correspond to the Watershed Boundary Dataset (WBD) as published by
the U.S. Natural Resources and Conservation Service and U.S. Geological Survey
8/29/2013 and within National Hydrography Dataset (NHD) version NHDH_OR_931v220.
}
\details{
\itemize{
\item HUC8: Eight digit subbasin code
\item HUC8_Name: Subbasin name
\item State: States within the HUC
}
}
\examples{
huc8 <- odeqmloctools::orhuc8
}
\keyword{datasets}
