% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orwqs.R
\docType{data}
\name{orwqs}
\alias{orwqs}
\title{Oregon Beneficial Use and Water Quality Standard Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1516986 rows and 12 columns.
}
\usage{
data(orwqs)
}
\description{
Table of Oregon Beneficial Use and Water Quality Standard Codes as of 11-10-2022.
The codes are referenced to National Hydrography Dataset (NHD) Permanent
Identifiers. The NHD version corresponds to NHDH_OR_931v220, which is the
current version used for DEQ business data.
}
\details{
See Oregon DEQ Georeferencing Classification Codes for meaning of Beneficial
Use and Water Quality Standard Codes.

\itemize{
\item Permanent_Identifier: NHD Permanent Identifier
\item WBArea_Permanent_Identifier: NHD Waterbody feature Permanent Identifier
\item FishCode: Designated Aquatic Life Fish Use Subcategories (Temperature)
\item SpawnCode: Designated Seasonal Spawning Date Ranges (Temperature)
\item WaterTypeCode: Water Type - Salinity-based water type designation
\item WaterBodyCode: Water Body Type - Functional or morphologic water body type designation
\item BacteriaCode: Designated Contact Recreation and Shellfish Harvest Use Subcategories (Bacteria)
\item DO_code: Designated Aquatic Life Use Subcategories (Dissolved Oxygen).
\item ben_use_code: Beneficial Use Designations – Designated beneficial uses of the waterbody
\item pH_code: Designated pH standards (pH)
\item DO_SpawnCode: Designated Seasonal Spawning Date Ranges (Dissolved Oxygen)
\item TempCode: Temperature standard applied for Clean Water Act purposes.
}
}
\examples{
wqs <- odeqmloctools::orwqs
}
\keyword{datasets}
