% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county.R
\name{get_county}
\alias{get_county}
\title{Get county}
\usage{
get_county(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs].
Typically entered using the numeric EPSG value. Accepts a vector.}
}
\value{
string of county basename
}
\description{
The function will query the Untied States Census Bureau states and county
TIGERweb feature service to determine the county name. The x and y
coordinates (longitude and latitude) are used to select a specific county.
}
\details{
The feature service can be accessed at https://tigerweb.geo.census.gov/arcgis/rest/services/TIGERweb/State_County/MapServer/13
}
