% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc8.R
\name{get_huc8code}
\alias{get_huc8code}
\title{Get HUC8 code}
\usage{
get_huc8code(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as
\code{\link[sf:st_crs]{sf::st_crs}}. Typically entered using
the numeric EPSG value. Accepts a vector.}
}
\value{
The HUC8 code as character format
}
\description{
The function will query Oregon DEQ's WBD feature service to determine the
HUC8 subbasin code. The x and y coordinates (longitude and latitude) are
used to select a specific HUC8 from the feature service.
}
\details{
The feature service can be accessed at \url{https://arcgis.deq.state.or.us/arcgis/rest/services/WQ/WBD/MapServer/1}.
}
\seealso{
\code{\link{get_huc8}}, \code{\link{get_huc8name}}
}
