% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhd_df.R
\name{get_nhd_df}
\alias{get_nhd_df}
\title{Get NHD High info}
\usage{
get_nhd_df(x, y, crs, search_dist = 100, service = "DEQ")
}
\arguments{
\item{x}{The longitude in decimal degrees. Required. Accepts a vector.}

\item{y}{The latitude in decimal degrees. Required. Accepts a vector.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs].
Typically entered using the numeric EPSG value. Accepts a vector.}

\item{search_dist}{The maximum search distance around x and y to look for features.
Measured in meters. Default is 100.}

\item{service}{The feature service to query. Options include "DEQ" or "USGS". Default is "DEQ".
If USGS, the query will be send to the NHD flowline feature service on the National Map.}
}
\value{
sf object with data frame columns for Permanent_Identifier, ReachCode, measure, and Snap_Distance
}
\description{
The function will query the NHD High REST service from Oregon DEQ or USGS and return a
dataframe containing the NHD flowline info. The supplied x and y coordinates
(longitude and latitude) are used to select the closest NHD flowline record. Only the closest
flowline within the search distance is returned. If there are no flowlines
within the search distance the returned dataframe will contain all NAs. If two or more flowline
records are equal distance to x and y only the first record will be returned.
}
\details{
Two REST feature services can be accessed. DEQ and USGS. Use the 'service'
argument to specific which service to use. DEQ's REST service is based
on NHDH_OR_931v220 at https://arcgis.deq.state.or.us/arcgis/rest/services/WQ/NHDH_ORDEQ/MapServer/1.
USGS's national map REST service is the most recent version of NHD high and
located at https://hydro.nationalmap.gov/arcgis/rest/services/nhd/MapServer/4.
}
