% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state.R
\name{get_state}
\alias{get_state}
\title{Get state abbreviation}
\usage{
get_state(x, y, crs)
}
\arguments{
\item{x}{The longitude in decimal degrees.}

\item{y}{The latitude in decimal degrees.}

\item{crs}{The coordinate reference system for x and y. Same format as [sf::st_crs].
Typically entered using the numeric EPSG value. Accepts a vector.}
}
\value{
state code
}
\description{
The function will query the Untied States Census Bureau states and county
TIGERweb feature service to determine the state abbreviation. The x and y
coordinates (longitude and latitude) are used to determine the state.
}
\details{
The feature service can be accessed at https://tigerweb.geo.census.gov/arcgis/rest/services/TIGERweb/State_County/MapServer/14
}
