% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_map.R
\name{launch_map}
\alias{launch_map}
\title{Launch a web map in a browser to review and edit monitoring location information}
\usage{
launch_map(mloc, px_ht = 470)
}
\arguments{
\item{mloc}{Data frame of the monitoring location information.
Required columns are 'Monitoring.Location.ID', 'Monitoring.Location.Name',
'Latitude', and 'Longitude'. Other columns are optional but will be added
if not present and the values set to NA. The full set of mloc column names
can be generated using \code{\link[odeqcdr:cols_mloc]{odeqcdr::cols_mloc}}
or as a data frame using \code{\link{df_mloc}}. 'Snap.Lat' and 'Snap.Long'
are added.}

\item{px_ht}{Height of the map in pixels. Default is 470 which fits on most standard laptop screens. The minimum height is 300 pixels.}
}
\value{
Launches a leaflet map within a Shiny app. Returns mloc data frame with any saved changes on app close.
}
\description{
Launch a web map in a browser to review and edit monitoring location information
}
