% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplus_na.R
\name{nhdplushr_na}
\alias{nhdplushr_na}
\title{NHDplus HR NA data frame}
\usage{
nhdplushr_na()
}
\description{
Return a data frame with all the columns in the NHDplus HR REST service with
NAs. Columns 'OBJECTID', 'Shape', and 'Shape_Length' are not included.
Columns 'Measure', 'Snap_Lat', 'Snap_Long', and 'Snap_Distance' are added for
downstream binding.
}
\details{
More info at at \url{https://hydro.nationalmap.gov/arcgis/rest/services/NHDPlus_HR/MapServer/2}.
}
