% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orau.R
\docType{data}
\name{orau}
\alias{orau}
\title{Oregon DEQ Assessment Units}
\format{
An object of class \code{data.frame} with 6182 rows and 15 columns.
}
\usage{
data(orau)
}
\description{
Table of Oregon DEQ Assessment Units. The table corresponds
to the final assessment units published with the 2022 Integrated Report.
Version 03-30-2022.
}
\details{
\itemize{
\item AU_ID: Oregon assessment unit ID.
\item AU_Name: Name of the assessment unit.
\item AU_WBType: Assessment unit waterbody type code.
\item AU_UseCode: Assessment unit use code.
\item AQWMS_NUM: Unique ID formatted as numeric.
\item AQWMS_TXT: Unique ID formatted as text.
\item AU_Description: Assessment unit descriptions.
\item AU_LenMile: Assessment unit length in miles.
\item AU_AreaAcre: Assessment unit areas in acres if a waterbody (e.g. lake).
\item HUC6: Six digit basin code where the assessment unit is located.
\item HUC6_Name: Basin name where the Assessment unit is located
\item HUC8: Eight digit subbasin code where the Assessment unit is located
\item HUC8_Name: Subbasin name where the assessment unit is located.
\item HUC10: Ten digit watershed code where the assessment unit is located.
\item HUC10_Name: Watershed name where the assessment unit is located.
}
}
\examples{
au <- odeqmloctools::orau

}
\keyword{datasets}
