% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orhuc12.R
\docType{data}
\name{orhuc12}
\alias{orhuc12}
\title{Oregon HUC12 subwatersheds}
\format{
An object of class \code{data.frame} with 3127 rows and 3 columns.
}
\usage{
data(orwhuc12)
}
\description{
Table of Oregon twelve digit hydrological codes and names.
Codes and names correspond to the Watershed Boundary Dataset (WBD) as published by
the U.S. Natural Resources and Conservation Service and U.S. Geological Survey
8/29/2013 and within National Hydrography Dataset (NHD) version NHDH_OR_931v220.
}
\details{
\itemize{
\item HUC12: Twelve digit subwatershed code
\item HUC12_Name: Subwatershed name
\item State: States within the HUC
}
}
\examples{
huc12 <- odeqmloctools::orhuc12
}
\keyword{datasets}
